requirements.txt
paho-mqtt==1.2.3
pyserial==3.3
